style: code cleaning
